lower versions compatible.
This update is fully compatible from 1.9.x to 1.15.x. In MinecraftP 0.32
(the previous version), only 1.12 behaved differently when placing a
signboard obtained via the /want cp command or /want creator. However,
in this version, it can be used just by placing it on 1.12, and the
support has been expanded from the lower version 1.9.x to 1.11.x.
MinecraftP ver 0.33 supports vanilla versions from 1.9.x to 1.15.x.